tests: make prefix truncate test more robust
The original intent of this test was to verify byte-for-byte equivalence
of a recovered replica from partitions that had undergone garbage
collection due to retention policy.

Since retention policy is applied independently on each node, this
turned out to be error prone / difficult to do. This patch rewrites the
test to verify higher level properties regarding under replication and
resulting offsets on the recovered replica.

Fixes: #2460

Signed-off-by: Noah Watkins <noah@vectorized.io>